refactor QuizProgressBar2 to QuizProgressBar
